+++ serial/pwm.h
@@ -1,7 +1,7 @@
-const double MAX_PWM = 0b11111111;
-const double NORM_PWM = 1 / MAX_PWM;
+const float MAX_PWM = 0b11111111;
+const float NORM_PWM = 1 / MAX_PWM;
 
-void commandToPwm(double norm, uint8_t& lpwm, uint8_t& rpwm)
+void commandToPwm(float norm, uint8_t& lpwm, uint8_t& rpwm)
 {
   if (norm < 0.0)
   {
@@ -17,7 +17,7 @@ void commandToPwm(double norm, uint8_t& lpwm, uint8_t& rpwm)
   }
 }
 
-double pwmToCommand(uint8_t lpwm, uint8_t rpwm)
+float pwmToCommand(uint8_t lpwm, uint8_t rpwm)
 {
-  return double(lpwm) * NORM_PWM - double(rpwm) * NORM_PWM;
+  return float(lpwm) * NORM_PWM - float(rpwm) * NORM_PWM;
 }
